Django/django-templates/fullstack/requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
django-stubs>=4.2.0

# Documentation
Sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
